fix: resolve 262 compilation errors in Wave 3 xpath-engine conversion
Fix readExternal signature mismatch (PrototypeFactory? -> PrototypeFactory),
companion object constant/method references, private field access across
Kotlin files, nullable type mismatches, array type invariance, and API
call patterns (size() -> size, replaceAll -> replace(Regex)).

+++ commcare-core/src/main/java/org/javarosa/xpath/expr/XPathIdCompressFunc.kt
@@ -19,7 +19,7 @@ open class XPathIdCompressFunc : XPathFuncExpr {
         model: DataInstance<*>, evalContext: EvaluationContext,
         evaluatedArgs: Array<Any?>
     ): Any {
-        val input = FunctionUtils.toInt(evaluatedArgs[0]).toInt()
+        val input = FunctionUtils.toInt(evaluatedArgs[0]).toLong()
         val growthDigits = FunctionUtils.toString(evaluatedArgs[1])
         val leadDigits = FunctionUtils.toString(evaluatedArgs[2])
         val bodyDigits = FunctionUtils.toString(evaluatedArgs[3])
